backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
yt-dlp>=2024.08.06
google-cloud-storage
google-cloud-aiplatform
google-generativeai
supabase
pydantic
moviepy
pytubefix
